+++ src/SC_test2.cpp
@@ -28,6 +28,7 @@
 #include "StrModule.h"
 
 #include <boost/test/unit_test.hpp>
+#include <cstdint>
 
 BOOST_AUTO_TEST_SUITE(SC_test)
 
@@ -435,7 +436,7 @@ BOOST_AUTO_TEST_CASE(Nondeterminism_detection){
   Configuration conf = DPORDriver_test::get_sc_conf();
   char changing = 0;
   std::stringstream ss;
-  ss << "inttoptr(i64 " << (unsigned long)(&changing) << " to i8*)";
+  ss << "inttoptr(i64 " << (uintptr_t)(&changing) << " to i8*)";
   std::string changingAddr = ss.str();
   std::string module = StrModule::portasm(R"(
 @x = global i32 0, align 4
@@ -666,7 +667,7 @@ BOOST_AUTO_TEST_CASE(Nondeterminism_without_branch){
   Configuration conf = DPORDriver_test::get_sc_conf();
   char changing = 0;
   std::stringstream ss;
-  ss << "inttoptr(i64 " << (unsigned long)(&changing) << " to i8*)";
+  ss << "inttoptr(i64 " << (uintptr_t)(&changing) << " to i8*)";
   std::string changingAddr = ss.str();
   std::string module = StrModule::portasm(R"(
 @x = global i32 0, align 4
